use a main.swift instead of @main attribute
