+++ Swish/Sources/appstore/AppStore.swift
@@ -1,32 +1,9 @@
 import Foundation
 import Sh
 import ShXcrun
-import ShGit
 import Rainbow
 
-@main
 struct AppStore {
-  static func main() {
-
-    do {
-
-      let git = Git()
-      guard try git.isClean()
-      else {
-        throw Errors.gitRepoNotClean
-      }
-
-      let logRoot = "Swish/logs"
-      let artifactRoot = "Swish/artifacts"
-      try sh(.terminal, "mkdir -p \(logRoot)")
-      try sh(.terminal, "mkdir -p \(artifactRoot)")
-
-      try AppStore(secrets: Secrets(), logRoot: logRoot, artifactRoot: artifactRoot).build()
-
-    } catch {
-      print("error".red, error)
-    }
-  }
 
   enum Errors: Error {
     case gitRepoNotClean
